Allow createauxblock RPC with different coinbase addresses
AuxpowMiner::createAuxBlock() only caches a single block for a specific
height. If we call createauxblock RPC multiple times with different
coinbase payout addresses in that case, they will actually all use the
first address. This commit uses a map with the coinbase transaction
script hash as the cache so that we can get the correct results.